requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.3
pathspec==0.8.1
pylint==2.6.0
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
